PROJ-72_privateLayout refactor: refactor imports for useMenuHandlers
